requirements.txt
Flask>=3.0.0
python-dotenv
mysql-connector-python
pandas
google-generativeai
scikit-learn 
joblib 
numpy